add parallel build capability
This allow all the compile jobs for a given executable to run in
parallel. The script will launch as many jobs as required up to a
maximum, which by default is the output of 'nproc', revert to a
hard-coded value of 4 if 'nproc' is not found. This can be overridden in
the settings.sh file or via the NJOBS environment variable.